# Speed up nvfp4 pack/unpack w/ torch.compile