@kylesayrs 2025-07-29
@fynnsu You should do at least a quick accuracy evaluation using lm_eval, comparing an [old model](https://huggingface.co/RedHatAI/phi-4-quantized.w4a16) to a newly packed model.

```python3
import argparse
import lm_eval
from lm_eval.utils import make_table

def main():
    results = lm_eval.simple_evaluate(
        model="hf",
        model_args={
            "pretrained": "MODEL_ID",
            "add_bos_token": True,
            "dtype": "auto",
        },
        tasks="arc_challenge_llama",
        batch_size=128,
        apply_chat_template=True,
        fewshot_as_multiturn=True,
    )

    print(make_table(results))
```