@kylesayrs 2025-07-29
Alternatively you can just check that the safetensor outputs are exactly the same, which is probably faster